+++ packages/terra-functional-testing/src/terra-cli/wdio/index.js
@@ -184,7 +184,7 @@ const cli = {
         describe: 'Whether or not to automatically update all reference screenshots with the latest screenshots.',
         default: false,
       },
-      useRemoteScreenshots: {
+      useRemoteReferenceScreenshots: {
         type: 'boolean',
         describe: 'A flag to download reference screenshots from a remote site for screenshot comparisons instead of using the local reference screenshots.',
         default: false,