+++ packages/terra-functional-testing/tests/jest/terra-cli/wdio/test-runner.test.js
@@ -54,7 +54,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         theme: 'terra-default-theme',
         updateScreenshots: true,
-        useRemoteScreenshots: true,
+        useRemoteReferenceScreenshots: true,
         useSeleniumStandaloneService: true,
       };
 
@@ -221,7 +221,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         themes: ['terra-default-theme'],
         updateScreenshots: true,
-        useRemoteScreenshots: true,
+        useRemoteReferenceScreenshots: true,
         useSeleniumStandaloneService: true,
       });
 
@@ -241,7 +241,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         theme: 'terra-default-theme',
         updateScreenshots: true,
-        useRemoteScreenshots: true,
+        useRemoteReferenceScreenshots: true,
         useSeleniumStandaloneService: true,
       });
     });