+++ packages/terra-functional-testing/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added `useRemoteScreenshots` cli option for downloading reference screenshots from a remote site for screenshot comparisons.
+  * Added `useRemoteReferenceScreenshots` cli option for downloading reference screenshots from a remote site for screenshot comparisons.
 
 ## 2.7.0 - (February 11, 2022)